+++ fearless-utils/src/main/java/jp/co/soramitsu/fearless_utils/encrypt/seed/ethereum/EthereumSeedFactory.kt
@@ -2,8 +2,8 @@ package jp.co.soramitsu.fearless_utils.encrypt.seed.ethereum
 
 import jp.co.soramitsu.fearless_utils.encrypt.mnemonic.Mnemonic
 import jp.co.soramitsu.fearless_utils.encrypt.mnemonic.MnemonicCreator
-import jp.co.soramitsu.fearless_utils.encrypt.mnemonic.SeedCreator
-import jp.co.soramitsu.fearless_utils.encrypt.mnemonic.SeedFactory
+import jp.co.soramitsu.fearless_utils.encrypt.seed.SeedCreator
+import jp.co.soramitsu.fearless_utils.encrypt.seed.SeedFactory
 
 object EthereumSeedFactory : SeedFactory {
